Make function to create grid
A function named createGrid was declared. This function will be passed
an argument called 'squares' that will define how many rows and columns
the grid is going to have.
This was made possible using a for loop for appehending the rows
inside the grid container, and placing another for loop inside that one
for appehending the columns inside the rows.