Create button for using random colors
A new boolean variable was initialized as false (UseRandomColor).
Then, a button was added to the DOM. This button will make use
of the variable that was just created.
An event listener was added to the button.
When the button is clicked, the UseRandomColor variable becomes true.
If the button were to be clicked again, the variable will become false.
Then, the text content of the button changes depending of wether
the variable is set to false or to true.

Then, inside the drawing for loop, the function to change the
background color of the cells has changed.

Now, if the UseRandomColor variable is true, the background color
set to the cells will be random (Using the function already declared).
If the UseRandomColor is false, then the color becomes black again.